AlgoCandlePredictor/requirements.txt
# Web Framework & API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Data Manipulation & Analysis  
pandas==2.1.4
numpy==1.24.4   # Compatible with TF 2.15.0

# Machine Learning - XGBoost & LSTM
xgboost==2.0.3
lightgbm==4.2.0
scikit-learn==1.4.0
tensorflow==2.15.0
keras==2.15.0

# Technical Indicators
ta==0.11.0

# Statistical Analysis & Kelly Criterion
scipy==1.11.4

# Model Persistence
joblib==1.3.2

# Configuration & Data
pyyaml==6.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
tqdm==4.66.1
